[OTS-694] (#65)
