+++ ios/OTAcceleratorPackUtilProject/OTAcceleratorPackUtil/OTCommonCommunicator.m
@@ -1,0 +1,13 @@
+//
+//  OTCommonCommunicator.m
+//  Pods
+//
+//  Created by Javier Fuchs on 12/21/16.
+//  Copyright © 2016 Tokbox, Inc. All rights reserved.
+//
+
+#import "OTCommonCommunicator.h"
+
+@implementation OTCommonCommunicator
+
+@end
