+++ ios/OTAcceleratorPackUtilProject/OTAcceleratorPackUtilProject/TestOneToOneCommunicatorSwitchingViewController.m
@@ -21,7 +21,7 @@
     
     self.communicator = [[OTOneToOneCommunicator alloc] init];
     self.communicator.dataSource = self;
-    [self.communicator connectWithHandler:^(OTOneToOneCommunicationSignal signal, NSError *error) {
+    [self.communicator connectWithHandler:^(OTCommunicationSignal signal, NSError *error) {
         if (signal == OTPublisherCreated && !error) {
             self.communicator.publisherView.frame = self.publisherView.bounds;
             [self.publisherView addSubview:self.communicator.publisherView];
